+++ CHANGELOG.md
@@ -15,6 +15,7 @@ See [changeset](https://github.com/wmontwe/mhp-mobile-challenge-android/compare/
 ### Added
 
 - App navigation using Compose Navigation
+- Compose test setup
 
 ## [0.0.1](https://github.com/wmontwe/mhp-mobile-challenge-android/releases/tag/v0.0.1)